AGR-19 fix en la table de merchants

+++ database/migrations/2024_12_24_153415_create_merchants_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('business_name');
             $table->string('trade_name')->nullable();
-            $table->integer('fiscal_number');
+            $table->bigInteger('fiscal_number');
             $table->string('main_activity')->nullable();
             $table->string('email');
             $table->string('phone');
